refactor: move ChatActionMessage component
